requirements.txt
absl-py==0.7.1
asgiref==3.2.3
astor==0.8.0
astroid==2.3.3
backcall==0.1.0
binarytree==4.1.0
bleach==3.0.2
colorama==0.4.3
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
Django==3.0.2
entrypoints==0.2.3
gast==0.2.2
grpcio==1.21.1
gunicorn==20.0.4
h5py==2.9.0
imageio==2.5.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.21
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.0
matlab==0.1
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
mysql==0.0.2
mysql-connector-python==8.0.14
mysqlclient==1.3.14
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.3
notebook==5.7.2
numpy==1.16.4
pandas==0.23.1
pandocfilters==1.4.2
parso==0.3.1
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.8.0
Pygments==2.3.0
pylint==2.4.4
pyparsing==2.3.0
python-dateutil==2.7.3
python2==1.2
pytz==2018.4
PyWavelets==1.0.3
pywinpty==0.5.4
PyYAML==5.1
pyzmq==17.1.2
qtconsole==4.4.3
scikit-image==0.15.0
scikit-learn==0.20.1
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
sqlparse==0.3.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
Theano==1.0.2
toolkits==0.1.28
torch==1.1.0
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.4.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
wrapt==1.11.2
xlrd==1.1.0
